refactor : dto 와 domain 분리
