+++ app/src/main/java/woowacourse/movie/domain/MovieInfoGetter.kt
@@ -1,23 +1,23 @@
 package woowacourse.movie.domain
 
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
 import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
-import kotlin.ranges.step
 
-@Parcelize
-data class MovieInfo(
-    val poster: Int,
-    val title: String,
-    val startDate: String,
-    val endDate: String,
-    val runningTime: Int,
-) : Parcelable {
+object MovieInfoGetter {
+    private const val DATE_PATTERN = "yyyy.M.d"
+
+    private const val WEEKEND_START_HOUR = 9
+    private const val WEEKDAY_START_HOUR = 10
+    private const val LAST_HOUR = 24
+    private const val HOUR_STEP = 2
+
     private val dateFormatter: DateTimeFormatter = DateTimeFormatter.ofPattern(DATE_PATTERN)
 
-    fun getDates(): List<String> {
+    fun getDates(
+        startDate: String,
+        endDate: String,
+    ): List<String> {
         val parsedStartDate = LocalDate.parse(startDate, dateFormatter)
         val parsedEndDate = LocalDate.parse(endDate, dateFormatter)
 
@@ -44,14 +44,5 @@ data class MovieInfo(
         return (startHour..LAST_HOUR step HOUR_STEP).map { hour ->
             String.format("%02d:00", hour)
         }
-    }
-
-    companion object {
-        private const val DATE_PATTERN = "yyyy.M.d"
-
-        private const val WEEKEND_START_HOUR = 9
-        private const val WEEKDAY_START_HOUR = 10
-        private const val LAST_HOUR = 24
-        private const val HOUR_STEP = 2
     }
 }
